LocalIpAddressLayoutRenderer - Moved internal interfaces to the NLog.Internal namespace
