+++ src/NLog/Internal/INetworkInterfaceRetriever.cs
@@ -33,7 +33,7 @@
 
 #if !NETSTANDARD1_3 && !NETSTANDARD1_5
 
-namespace NLog.LayoutRenderers
+namespace NLog.Internal
 {
     using System.Net.NetworkInformation;
