+++ src/NLog/Internal/NetworkInterfaceRetriever.cs
@@ -33,11 +33,10 @@
 
 #if !NETSTANDARD1_3 && !NETSTANDARD1_5
 
-using System.Collections.Generic;
-using System.Net.NetworkInformation;
-
-namespace NLog.LayoutRenderers
+namespace NLog.Internal
 {
+    using System.Net.NetworkInformation;
+
     /// <summary>
     /// Retrieve network interfaces
     /// </summary>
